Tools/tools/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
Pillow==10.4.0
PyPDF2==3.0.1
selenium==4.24.0
requests==2.32.3
webdriver-manager==4.0.2
bs4==0.0.2
